server/requirements.txt
fastapi==0.135.3
uvicorn[standard]
open-clip-torch==3.3.0
torch>=2.0.0
numpy>=2.0
plyfile>=0.9
python-dotenv>=1.0
scikit-learn>=1.8.0
scipy>=1.17.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.28.0
pytest>=8.0
pytest-asyncio>=0.24
